pyproject.toml
[project]
name = "readoutloud"
version = "0.1.0"
description = "Text to Speech tool built on streamlit and piper-tts"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "piper-tts==1.3.0",
    "sounddevice==0.5.2",
    "soundfile==0.13.1",
    "streamlit==1.50.0",

]
